requirements.txt
cdifflib==1.2.6
discord.py==2.2.3
fastapi==0.95.2
gradio==3.32.0
httpx==0.24.1
pandas==2.0.1
Pillow==9.5.0
pydantic==1.10.7
starlette==0.27.0
uvicorn==0.22.0
openai
tiktoken
